Move cowTabletList implementation to tablets.go file
